+++ saleor/csv/tests/test_notifications.py
@@ -12,9 +12,11 @@ from ..notifications import get_default_export_payload
 
 @freeze_time("2018-05-31 12:00:01")
 @mock.patch("saleor.plugins.manager.PluginsManager.gift_card_export_completed")
+@mock.patch("saleor.plugins.manager.PluginsManager.product_export_completed")
 @mock.patch("saleor.plugins.manager.PluginsManager.notify")
 def test_send_export_download_link_notification(
     mocked_notify,
+    mocked_product_export_completed,
     mocked_gift_card_export_completed,
     site_settings,
     user_export_file,
@@ -43,13 +45,16 @@ def test_send_export_download_link_notification(
         AdminNotifyEvent.CSV_EXPORT_SUCCESS, expected_payload
     )
     mocked_gift_card_export_completed.assert_not_called()
+    mocked_product_export_completed.assert_called_with(user_export_file)
 
 
 @freeze_time("2018-05-31 12:00:01")
 @mock.patch("saleor.plugins.manager.PluginsManager.gift_card_export_completed")
+@mock.patch("saleor.plugins.manager.PluginsManager.product_export_completed")
 @mock.patch("saleor.plugins.manager.PluginsManager.notify")
 def test_send_export_failed_info(
     mocked_notify,
+    mocked_product_export_completed,
     mocked_gift_card_export_completed,
     site_settings,
     user_export_file,
@@ -79,3 +84,4 @@ def test_send_export_failed_info(
         AdminNotifyEvent.CSV_EXPORT_FAILED, expected_payload
     )
     mocked_gift_card_export_completed.assert_called_once_with(user_export_file)
+    mocked_product_export_completed.assert_not_called()