+++ saleor/graphql/csv/mutations/export_products.py
@@ -80,6 +80,10 @@ class ExportProducts(BaseExportMutation):
                 type=WebhookEventAsyncType.NOTIFY_USER,
                 description="A notification for the exported file.",
             ),
+            WebhookEventInfo(
+                type=WebhookEventAsyncType.PRODUCT_EXPORT_COMPLETED,
+                description="A notification for the exported file.",
+            ),
         ]
 
     @classmethod